cloud-functions/smart-harvest/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.44.2
torch==2.4.1
Pillow==10.4.0
python-multipart==0.0.9
numpy==1.26.4
pytest==8.3.3
httpx==0.27.2
